# src/main/java/org/bitebuilders/controller/dto/ReferralTokenDTO.java
package org.bitebuilders.controller.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.ToString;

@Data
@AllArgsConstructor
@ToString
public class ReferralTokenDTO {
    private String referralToken;
}