# src/main/resources/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CRM module for Uralintern",
    "description": "API documentation for Uralintern CRM Module.\n\n**Контроллеры:**\n- **auth-controller**: Отвечает за регистрацию и авторизацию пользователей, включая регистрацию с реферальной ссылкой и обновление токенов.\n- **user-info-controller**: Управляет пользовательскими данными, включая обновление профиля, получение реферальных ссылок и список руководителей.\n- **event-controller**: Обеспечивает управление мероприятиями, включая их создание, обновление, скрытие, удаление, а также доступ к информации о мероприятиях.\n- **message-controller**: Управляет шаблонами сообщений для мероприятий, включая их получение и обновление.\n- **event-student-controller**: Обеспечивает управление заявками и участием студентов в мероприятиях, включая отправку заявок, их принятие/отклонение и управление студентами на мероприятии.\n- **event-curator-controller**: Управляет заявками и участием кураторов в мероприятиях, включая их принятие/отклонение и управление кураторами.\n- **notification-controller**: Управляет уведомлениями пользователей, связанными с их участием в мероприятиях.\n",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://localhost/api",
      "description": "kayakto server url"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/users/invite-admin": {
      "post": {
        "tags": [
          "user-info-controller"
        ],
        "operationId": "inviteManager",
        "description": "Создание реферальной ссылки для приглашения администратора. Доступно только администраторам.\n",
        "responses": {
          "200": {
            "description": "Реферальная ссылка успешно создана.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferralTokenDTO"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/users/update-password": {
      "post": {
        "tags": [
          "user-info-controller"
        ],
        "operationId": "updatePassword",
        "description": "Обновление пароля пользователя, который сделал запрос. Доступно всем пользователям.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokensDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/users/update-email": {
      "post": {
        "tags": [
          "user-info-controller"
        ],
        "operationId": "updateEmail",
        "description": "Обновление почты пользователя, который сделал запрос. Доступно всем пользователям.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokensDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": [
          "user-info-controller"
        ],
        "operationId": "getCurrentUser",
        "description": "Получение данных текущего пользователя.\nТребуется авторизация.\n",
        "responses": {
          "200": {
            "description": "Данные пользователя успешно получены.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      },
      "put": {
        "tags": [
          "user-info-controller"
        ],
        "operationId": "updateCurrentUser",
        "description": "Обновление данных текущего пользователя.\nТребуется авторизация.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Данные пользователя успешно обновлены.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/users/my-role": {
      "get": {
        "tags": [
          "user-info-controller"
        ],
        "operationId": "getMyRole",
        "description": "Получение роли текущего пользователя. Доступно всем пользователям. Вернет роль в виде \"STUDENT\"",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": [
          "user-info-controller"
        ],
        "operationId": "getUserById",
        "description": "Получение профиля пользователя по идентификатору. Доступ - все авторизованные пользователи. В будущем может будет ограничение доступа.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
...
      "BadRequestResponse": {
        "description": "Bad Request, запрос неверный со стороны клиента",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Invalid argument"
                }
              }
            }
          }
        }
      },
      "ForbiddenResponse": {
        "description": "Forbidden, отказано в доступе",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Access denied"
                }
              }
            }
          }
        }
      },
      "NotFoundResponse": {
        "description": "Not Found, такого элемента не существует",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Event with id 123 not found"
                }
              }
            }
          }
        }
      },
      "InternalServerErrorResponse": {
        "description": "Internal Server Error, запрос вызвал ошибку на сервере",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "UserUpdateRequest": {
        "required": [
          "firstName",
          "lastName",
          "telegramUrl",
          "vkUrl"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Николай"
          },
          "lastName": {
            "type": "string",
            "example": "Путин"
          },
          "surname": {
            "type": "string",
            "example": "Евгеньевич"
          },
          "telegramUrl": {
            "type": "string",
            "example": "https://t.me/kayakto"
          },
          "vkUrl": {
            "type": "string",
            "example": "https://vk.com/kayakto"
          },
          "competencies": {
            "type": "string",
            "example": "Spring Boot, Gradle, git, Java, Docker"
          }
        }
      },
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 123
          },
          "firstName": {
            "type": "string",
            "example": "Николай"
          },
          "lastName": {
            "type": "string",
            "example": "Иванов"
          },
          "surname": {
            "type": "string",
            "example": "Евгеньевич"
          },
          "telegramUrl": {
            "type": "string",
            "example": "https://t.me/kayakto"
          },
          "vkUrl": {
            "type": "string",
            "example": "https://vk.com/kayakto"
          },
          "role_enum": {
            "type": "string",
            "enum": [
              "ADMIN",
              "CURATOR",
              "MANAGER",
              "STUDENT"
            ],
            "example": "STUDENT"
          },
          "competencies": {
            "type": "string",
            "example": "Spring Boot, Gradle, git, Java, Docker"
          }
        }
      },
...
      "TokensDTO": {
        "type": "object",
        "properties": {
          "access": {
            "type": "string"
          },
          "refresh": {
            "type": "string"
          }
        }
      },
      "ReferralTokenDTO": {
        "type": "object",
        "properties": {
          "referralToken": {
            "type": "string"
          }
        }
      },
...
      "MessageResponseDTO": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "EmailUpdateRequest": {
        "type": "object",
        "properties": {
          "oldEmail": {
            "type": "string"
          },
          "newEmail": {
            "type": "string"
          }
        }
      },
      "PasswordUpdateRequest": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}